+++ ci_environment/android-sdk/attributes/default.rb
@@ -5,15 +5,13 @@ default['android-sdk']['owner']          = node['travis_build_environment']['use
 default['android-sdk']['group']          = node['travis_build_environment']['group']
 default['android-sdk']['setup_root']     = nil  # ark defaults (/usr/local) is used if this attribute is not defined
 
-default['android-sdk']['version']        = '22.6.2'
-default['android-sdk']['checksum']       = 'f87e1bb207783a8b7d59775dcd3a8a2828034787860e44c99a9aefebdf7cb28f'
+default['android-sdk']['version']        = '23'
+default['android-sdk']['checksum']       = '2eaab06852ac21b6c79df73c07a667c5da5be57f7ffcbd4f17aef7efeea22ac1'
 default['android-sdk']['download_url']   = "http://dl.google.com/android/android-sdk_r#{node['android-sdk']['version']}-linux.tgz"
 
 #
 # List of Android SDK components to preinstall:
-# Selection based on
-# - Platform usage statistics (see http://developer.android.com/about/dashboards/index.html)
-# - Build Tools releases: http://developer.android.com/tools/revisions/build-tools.html
+# TODO: deprecate notice
 #
 # Hint:
 # Add 'tools' to the list below if you wish to get the latest version,
@@ -21,21 +19,27 @@ default['android-sdk']['download_url']   = "http://dl.google.com/android/android
 # Note that it will require (waste) some extra download effort.
 default['android-sdk']['components']     = %w(platform-tools
                                               android-19
-                                              sysimg-19
+                                              sys-img-armeabi-v7a-android-19
                                               android-18
-                                              sysimg-18
+                                              sys-img-armeabi-v7a-android-18
                                               android-17
-                                              sysimg-17
+                                              sys-img-armeabi-v7a-android-17
                                               android-16
-                                              sysimg-16
+                                              sys-img-armeabi-v7a-android-16
                                               android-15
-                                              sysimg-15
+                                              sys-img-armeabi-v7a-android-15
                                               android-10
                                               extra-android-support
                                               extra-google-google_play_services
                                               extra-google-m2repository
                                               extra-android-m2repository)
 
-default['android-sdk']['license']['white_list']     = %w(android-sdk-license-.+)
+default['android-sdk']['license']['white_list']     = %w(.+)
 default['android-sdk']['license']['black_list']     = []    # e.g. ['intel-.+', 'mips-.+', 'android-wear-sdk-license-.+']
 default['android-sdk']['license']['default_answer'] = 'n'   # 'y' or 'n' ('yes' or 'no')
+
+default['android-sdk']['scripts']['path']           = '/usr/local/bin'
+default['android-sdk']['scripts']['owner']          = node['android-sdk']['owner']
+default['android-sdk']['scripts']['group']          = node['android-sdk']['group']
+
+default['android-sdk']['maven-rescue']              = false