+++ ci_environment/android-sdk/CHANGELOG.md
@@ -4,6 +4,9 @@ CHANGELOG for Android-SDK cookbook
 v0.2.0 (upcoming)
 -------------------
 
+- Integrate by default with Android SDK 23.0.0
+- WIP: Optionally install Maven Android SDK Deployer [GH-14]
+- Deploy scripts for waiting on Emulator startup [GH-16]
 - Deploy scripts for non-interactive SDK setup/updates [GH-13]
 
 v0.1.1 (2014-04-01)